planning: refactored DpStOptimizer to PathTimeHeuristicOptimizer
