+++ modules/planning/scenarios/lane_follow/BUILD
@@ -42,7 +42,7 @@ cc_library(
         "//modules/planning/tasks/deciders/lane_change_decider",
         "//modules/planning/tasks/optimizers:path_optimizer",
         "//modules/planning/tasks/optimizers:speed_optimizer",
-        "//modules/planning/tasks/optimizers/dp_st_speed:dp_st_speed_optimizer",
+        "//modules/planning/tasks/optimizers/path_time_heuristic:path_time_heuristic_optimizer",
         "//modules/planning/tasks/optimizers/path_decider",
         "//modules/planning/tasks/optimizers/speed_decider",
         "@eigen",
@@ -80,7 +80,7 @@ cc_library(
         "//modules/planning/scenarios:scenario",
         "//modules/planning/tasks/optimizers:path_optimizer",
         "//modules/planning/tasks/optimizers:speed_optimizer",
-        "//modules/planning/tasks/optimizers/dp_st_speed:dp_st_speed_optimizer",
+        "//modules/planning/tasks/optimizers/path_time_heuristic:path_time_heuristic_optimizer",
         "//modules/planning/tasks/optimizers/path_decider",
         "//modules/planning/tasks/optimizers/speed_decider",
         "@eigen",
