+++ modules/planning/tasks/optimizers/path_time_heuristic/dp_st_graph.h
@@ -35,8 +35,8 @@
 #include "modules/planning/common/speed/speed_data.h"
 #include "modules/planning/common/speed/st_point.h"
 #include "modules/planning/common/st_graph_data.h"
-#include "modules/planning/tasks/optimizers/dp_st_speed/dp_st_cost.h"
-#include "modules/planning/tasks/optimizers/dp_st_speed/st_graph_point.h"
+#include "modules/planning/tasks/optimizers/path_time_heuristic/dp_st_cost.h"
+#include "modules/planning/tasks/optimizers/path_time_heuristic/st_graph_point.h"
 
 namespace apollo {
 namespace planning {
