+++ modules/planning/tasks/optimizers/path_time_heuristic/dp_st_graph_test.cc
@@ -17,7 +17,7 @@
 /**
  * @file
  **/
-#include "modules/planning/tasks/optimizers/dp_st_speed/dp_st_graph.h"
+#include "modules/planning/tasks/optimizers/path_time_heuristic/dp_st_graph.h"
 
 #include "gtest/gtest.h"
