+++ modules/planning/tasks/task_factory.cc
@@ -32,7 +32,6 @@
 #include "modules/planning/tasks/deciders/path_bounds_decider/path_bounds_decider.h"
 #include "modules/planning/tasks/deciders/path_lane_borrow_decider/path_lane_borrow_decider.h"
 #include "modules/planning/tasks/deciders/speed_bounds_decider/speed_bounds_decider.h"
-#include "modules/planning/tasks/optimizers/dp_st_speed/dp_st_speed_optimizer.h"
 #include "modules/planning/tasks/optimizers/open_space_trajectory_generation/open_space_trajectory_provider.h"
 #include "modules/planning/tasks/optimizers/open_space_trajectory_partition/open_space_trajectory_partition.h"
 #include "modules/planning/tasks/optimizers/path_decider/path_decider.h"
@@ -42,6 +41,7 @@
 #include "modules/planning/tasks/optimizers/speed_decider/speed_decider.h"
 #include "modules/planning/tasks/rss/decider_rss.h"
 #include "modules/planning/tasks/task.h"
+#include "optimizers/path_time_heuristic/path_time_heuristic_optimizer.h"
 
 namespace apollo {
 namespace planning {
@@ -82,7 +82,7 @@ void TaskFactory::Init(const PlanningConfig& config) {
                          });
   task_factory_.Register(TaskConfig::DP_ST_SPEED_OPTIMIZER,
                          [](const TaskConfig& config) -> Task* {
-                           return new DpStSpeedOptimizer(config);
+                           return new PathTimeHeuristicOptimizer(config);
                          });
   task_factory_.Register(TaskConfig::PATH_DECIDER,
                          [](const TaskConfig& config) -> Task* {
